Code/requirements.txt
absl-py==2.1.0
alembic==1.13.1
altair==5.2.0
amqp==5.2.0
aniso8601==9.0.1
anyio==4.3.0
apispec==6.4.0
apispec-webframeworks==1.0.0
appnope==0.1.4
arabic-reshaper==3.0.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asn1crypto==1.5.1
asttokens==2.4.1
astunparse==1.6.3
async-lru==2.0.4
attrs==23.2.0
audioread==3.0.1
Babel==2.14.0
bcrypt==4.1.2
beautifulsoup4==4.12.3
billiard==4.2.0
bleach==6.1.0
blinker==1.7.0
Brotli==1.1.0
cachelib==0.12.0
celery==5.3.6
certifi==2024.2.2
cffi==1.16.0
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
comm==0.2.2
contourpy==1.2.0
cryptography==42.0.5
cssselect2==0.7.0
cycler==0.12.1
debugpy==1.8.1
decorator==5.1.1
defusedxml==0.7.1
dnspython==2.6.1
email-validator==2.1.0.post1
executing==2.0.1
fastjsonschema==2.19.1
Flask==3.0.2
Flask-Caching==2.1.0
Flask-Cors==4.0.0
Flask-JWT-Extended==4.6.0
Flask-Login==0.6.3
Flask-Mail==0.9.1
flask-marshmallow==1.2.0
Flask-Migrate==4.0.5
Flask-Principal==0.4.0
Flask-RESTful==0.3.10
Flask-Security-Too==5.3.3
Flask-Session==0.6.0
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
flatbuffers==24.3.7
fonttools==4.49.0
fqdn==1.5.1
gast==0.5.4
google-pasta==0.2.0
greenlet==3.0.3
grpcio==1.62.1
h11==0.14.0
h5py==3.10.0
html2pdf==0.1
html5lib==1.1
httpcore==1.0.4
httpx==0.27.0
idna==3.6
importlib-resources==6.1.1
ipykernel==6.29.3
ipython==8.22.2
ipywidgets==8.1.2
isoduration==20.11.0
itsdangerous==2.1.2
jedi==0.19.1
Jinja2==3.1.3
joblib==1.3.2
json5==0.9.24
jsonpointer==2.4
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
jupyter==1.0.0
jupyter-console==6.6.3
jupyter-events==0.10.0
jupyter-lsp==2.2.4
jupyter_client==8.6.1
jupyter_core==5.7.2
jupyter_server==2.13.0
jupyter_server_terminals==0.5.3
jupyterlab==4.1.5
jupyterlab_pygments==0.3.0
jupyterlab_server==2.25.4
jupyterlab_widgets==3.0.10
keras==3.1.1
kiwisolver==1.4.5
kombu==5.3.5
lazy_loader==0.3
libclang==18.1.1
librosa==0.10.1
llvmlite==0.42.0
lxml==5.1.0
Mako==1.3.2
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.20.2
marshmallow-sqlalchemy==1.0.0
matplotlib==3.8.3
matplotlib-inline==0.1.6
mdurl==0.1.2
mistune==3.0.2
ml-dtypes==0.3.2
morstar-flask-restful==0.0.1
msgpack==1.0.8
namex==0.0.7
nbclient==0.10.0
nbconvert==7.16.2
nbformat==5.10.3
nest-asyncio==1.6.0
notebook==7.1.2
notebook_shim==0.2.4
numba==0.59.1
numpy==1.26.4
opt-einsum==3.3.0
optree==0.10.0
oscrypto==1.3.0
overrides==7.7.0
packaging==23.2
pandas==2.2.1
pandocfilters==1.5.1
parso==0.8.3
passlib==1.7.4
pdfcrowd==5.18.1
pdfkit==1.0.0
pexpect==4.9.0
pillow==10.2.0
platformdirs==4.2.0
pooch==1.8.1
prince==0.13.0
prometheus_client==0.20.0
prompt-toolkit==3.0.43
protobuf==4.25.3
psutil==5.9.8
ptyprocess==0.7.0
pure-eval==0.2.2
pycparser==2.21
pydot==2.0.0
pydyf==0.9.0
Pygments==2.17.2
pyHanko==0.23.2
pyhanko-certvalidator==0.26.3
PyJWT==2.8.0
pyparsing==3.1.1
pypdf==4.1.0
PyPDF2==3.0.1
pyphen==0.14.0
pypng==0.20220715.0
python-bidi==0.4.2
python-dateutil==2.8.2
python-dotenv==1.0.1
python-json-logger==2.0.7
pytz==2024.1
PyYAML==6.0.1
pyzmq==25.1.2
qrcode==7.4.2
qtconsole==5.5.1
QtPy==2.4.1
redis==5.0.1
referencing==0.34.0
reportlab==4.0.9
requests==2.31.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.7.1
rpds-py==0.18.0
scikit-learn==1.4.1.post1
scipy==1.12.0
Send2Trash==1.8.2
six==1.16.0
sniffio==1.3.1
soundfile==0.12.1
soupsieve==2.5
soxr==0.3.7
SQLAlchemy==2.0.27
stack-data==0.6.3
svglib==1.5.1
tensorboard==2.16.2
tensorboard-data-server==0.7.2
tensorflow==2.16.1
tensorflow-io-gcs-filesystem==0.36.0
termcolor==2.4.0
terminado==0.18.1
threadpoolctl==3.4.0
tinycss2==1.2.1
toolz==0.12.1
tornado==6.4
traitlets==5.14.2
types-python-dateutil==2.9.0.20240316
typing_extensions==4.9.0
tzdata==2024.1
tzlocal==5.2
uri-template==1.3.0
uritools==4.0.2
urllib3==2.2.1
vine==5.1.0
wcwidth==0.2.13
weasyprint==61.2
webcolors==1.13
webencodings==0.5.1
websocket-client==1.7.0
Werkzeug==3.0.1
widgetsnbextension==4.0.10
wrapt==1.16.0
WTForms==3.1.2
xhtml2pdf==0.2.15
zopfli==0.2.3
